Lexer: Rename `CharPointer` -> `FileVisitor`.
